tidied up all the routes files and fixed some import bugs
