+++ src/routes/__init__.py
@@ -1,0 +1,3 @@
+from . import main
+from . import api
+from . import admin
