+++ src/routes/admin/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+adminbp  =   Blueprint('admin', __name__)
+
+from . import *
