+++ src/routes/api/chat.py
@@ -1,0 +1,4 @@
+from flask import redirect, url_for
+
+def chatbot():
+    return redirect(url_for('main.index')) # return data
