+++ src/routes/main/dashboard.py
@@ -1,0 +1,10 @@
+from flask import render_template
+
+def dashboard():
+    """
+    Renders the dashboard page.
+
+    Returns:
+        The rendered template for the dashboard page.
+    """
+    return render_template('dashboard.html')
