+++ src/routes/main/register.py
@@ -1,0 +1,10 @@
+from flask import render_template
+
+def register():
+    """
+    Renders the register.html template.
+
+    Returns:
+        The rendered register.html template.
+    """
+    return render_template('register.html')
